feat: 게시물을 생성한 User가 게시물을 ID로 가져오는 기능, 게시물을 수정하는 기능, 게시물을 삭제하는 기눙 구현

+++ src/boards/boards.service.ts
@@ -13,8 +13,10 @@ export class BoardsService {
     private boardRepository: BoardRepository,
   ) {}
 
-  async getBoardById(id: number): Promise<Board> {
-    const found = await this.boardRepository.findOne({ where: { id } });
+  async getBoardById(id: number, user: User): Promise<Board> {
+    const found = await this.boardRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!found) throw new NotFoundException(`Can't find board with id ${id}`);
     return found;
   }
@@ -33,16 +35,20 @@ export class BoardsService {
     return this.boardRepository.createBoard(createBoardDto, user);
   }
 
-  async deleteBoard(id: number): Promise<void> {
-    const result = await this.boardRepository.delete(id);
+  async deleteBoard(id: number, user: User): Promise<void> {
+    const result = await this.boardRepository.delete({ id, userId: user.id });
     if (result.affected === 0)
       throw new NotFoundException(`Can't find Board with id ${id}`);
   }
 
-  async updateBoardStatus(id: number, status: BoardStatus): Promise<Board> {
-    const board = await this.getBoardById(id);
+  async updateBoardStatus(
+    id: number,
+    status: BoardStatus,
+    user: User,
+  ): Promise<Board> {
+    const board = await this.getBoardById(id, user);
     board.status = status;
-    await this.boardRepository.save(board);
+    await board.save();
     return board;
   }
